Added files to create a docker image and fixes/improvements to allow to run this inside a container
The configuration file now allows a "- interval:" parameter for each config set and the processing of these configsets has been fixed so that they are processed in the defined intervals. No interval means "run only once", which is useful if data is not updated at all or very very rarely.

In addition to the dockerfile, the requirements.txt file lists all the Python dependencies that need to be installed when building the image.