Revised version with lots of changes
Including a web based configuration editor. For this, the docker implementation needs to use a separate start script that starts both data2mqtt and the web server (based on Flask).

The commit includes all files required for the web UI implementation as well as the dependencies required for creating the docker image.